services/fpe-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ff3==1.0.2
pycryptodome==3.23.0
psycopg2-binary==2.9.10
hvac==2.3.0
pydantic==2.10.3
python-json-logger==3.2.1
